+++ liquid_babel/__init__.py
@@ -1,4 +1,1 @@
-# flake8: noqa
-# pylint: disable=useless-import-alias,missing-module-docstring
-
-__version__ = "1.0.0"
+__version__ = "1.0.1"