+++ liquid_babel/messages/__init__.py
@@ -1,7 +1,7 @@
 """Liquid Babel message extraction and translation."""
-from .extract import extract_liquid
 from .extract import extract_from_template
 from .extract import extract_from_templates
+from .extract import extract_liquid
 
 __all__ = (
     "extract_liquid",